+++ src/app/accounts/account-dialog/account-dialog.component.ts
@@ -1,7 +1,7 @@
 import { Component, Inject, OnInit } from "@angular/core";
 import { FormBuilder, FormGroup, AbstractControl } from "@angular/forms";
 import { MatDialogRef, MAT_DIALOG_DATA } from "@angular/material";
-import { Account } from "../../interfaces/account";
+import { Account } from "../../interfaces/account.interface";
 
 interface AccountDialogData {
   action: string;
