+++ src/app/services/storage/tags.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@angular/core";
-import { Tag } from "../../interfaces/tag";
+import { Tag } from "../../interfaces/tag.interface";
 import { DatabaseService } from "./database.service";
 import { AbstractService } from "./abstract.service";
