+++ src/app/services/storage/transactions.service.ts
@@ -3,7 +3,7 @@ import { Injectable } from "@angular/core";
 import { DatabaseService } from "./database.service";
 import { AbstractService } from "./abstract.service";
 
-import { Transaction } from "../../interfaces/transaction";
+import { Transaction } from "../../interfaces/transaction.interface";
 
 @Injectable()
 export class TransactionsService extends AbstractService<Transaction> {
