+++ src/app/tags/tag-dialog/tag-dialog.component.ts
@@ -1,7 +1,7 @@
 import { Component, Inject, OnInit } from "@angular/core";
 import { FormBuilder, FormGroup, AbstractControl } from "@angular/forms";
 import { MatDialogRef, MAT_DIALOG_DATA } from "@angular/material";
-import { Tag } from "../../interfaces/tag";
+import { Tag } from "../../interfaces/tag.interface";
 
 interface TagDialogData {
   action: string;
