+++ src/app/transactions/table-transactions/table-transactions.component.ts
@@ -8,8 +8,8 @@ import {
 } from "@angular/core";
 import { MatTableDataSource, MatSort } from "@angular/material";
 
-import { Account } from "../../interfaces/account";
-import { Tag } from "../../interfaces/tag";
+import { Account } from "../../interfaces/account.interface";
+import { Tag } from "../../interfaces/tag.interface";
 
 interface ContentData {
   id: string;
