+++ src/app/transactions/transaction-dialog/transaction-dialog.component.ts
@@ -3,9 +3,9 @@ import { FormBuilder, FormGroup } from "@angular/forms";
 import { DatePipe } from "@angular/common";
 import { MAT_DIALOG_DATA, MatDialogRef } from "@angular/material";
 
-import { Transaction } from "../../interfaces/transaction";
-import { Account } from "../../interfaces/account";
-import { Tag } from "../../interfaces/tag";
+import { Transaction } from "../../interfaces/transaction.interface";
+import { Account } from "../../interfaces/account.interface";
+import { Tag } from "../../interfaces/tag.interface";
 
 interface TransactionDialogData {
   action: string;
